Refactored Node components. Fix for updating argsDict with default values for nodes with BooleanInputs.
