+++ nodestudiojs/src/components/graph/Node.scss
@@ -43,85 +43,4 @@ $node-canvas-bg-color: #e6e6e6;
       border-top-left-radius: 4px; 
       border-top-right-radius: 4px;
     }
-  
-    .node_io {
-      padding: 0 10px;
-      font-size: 12px;
-  
-      border-style: solid;
-      border-bottom-color: $grey-border-color;
-      border-width: 0px 0px 1px 0px;
-    }
-  
-    .node_io-item {
-      padding: 10px 0;
-  
-      &:last-child {
-          margin-bottom: 0px;
-      }
-    } 
-  
-    .node_io-point {
-      width: 12px;
-      height: 12px;
-      border-radius: 100px;
-      border-style: solid;
-      border-bottom-color: $grey-border-color;
-      border-width: 1px;
-    }
-  
-    .node_io-text {
-      margin-left: 6px;
-      margin-right: 6px;
-    }
-  
-    .node_output {
-      text-align: right;
-    }
-  
-    .node_input {
-      text-align: left;
-    }
-
-    .node_props {
-      width: 100%;
-      padding: 0.6em;
-      text-align: left;
-      background-color: $node-canvas-bg-color;
-      border-style: solid;
-      border-bottom-color: $grey-border-color;
-      border-width: 0px 0px 1px 0px;
-      
-      input {
-        background-color: lighten($app-background-color, 80%);
-        width: 100%;
-        font-size: 0.75em;
-      }
-  
-      label {
-        font-size: 0.75em;
-        font-weight: 200;
-        text-transform: capitalize;
-        margin: 0.5em 0 0.5em 0;
-      }
-
-      [type=file] {
-        margin-top: 1em;
-        margin-bottom: 0.5em;
-      }
-
-      button {
-        font-size: 0.75em;
-        padding: 0.5em;
-      }
-
-      .select {
-        margin-bottom: 0.5em;
-      }
-    }
-  
-    .node_image {
-      padding: 10px;
-      background-color: $node-canvas-bg-color;
-    }
-  }
+}
